Staging (#2)
* init

* update file css

* new css

* add document

* update style

* update

* add daily bible modal

* update sort import

* update export

* add integrate with firebase

* update package.json

* update params

* fix issue deploy

* update css

* update main width

* update css

* add layout

* Add game layout

* update new code

* add page game

* update some css

* add CuuUoc

* add UI

* Add read bible modal

* update css

* update css sign -in

* update signin

* update css index

* update css

* update css

* update css

* update css index

* update css index

* update css daily bible

* update css sign-in

* update css Game Home

* update css game Cuu Uoc

* update css game home

* update code for read bible of stage

* update css for mobile

* Add primary button (#2)

* Add primary button

* Full width

* feat: update page ui system

* Text box (#3)

* Custom text box & modify primary btn

* Custom text box & modify primary btn

* add icon text box

* add icon text box

* add icon text box

* fix: login function (#4)

* fix/user-login (#5)

* feat: apply new template (#1)


---------

+++ apps/hojo-web/components/protected/protected.tsx
@@ -1,9 +1,9 @@
-import { useRouter } from 'next/router';
+import { useRouter } from "next/router";
 
-import { FC, ReactElement, ReactNode } from 'react';
+import { FC, ReactElement, ReactNode } from "react";
 
-import { useBrowserLayoutEffect } from '@web/utils/hooks/useBrowserLayoutEffect';
-import mutateStorage from '@web/utils/mutate-storage';
+import { useBrowserLayoutEffect } from "@web/utils/hooks/useBrowserLayoutEffect";
+import mutateStorage from "@web/utils/mutate-storage";
 
 interface ProtectedProps {
   children: (ReactNode & ReactElement) | any;
@@ -11,17 +11,18 @@ interface ProtectedProps {
 
 export const Protected: FC<ProtectedProps> = ({ children }) => {
   const router = useRouter();
+  const pathWithoutProtected = ["/ui-system"];
 
   useBrowserLayoutEffect(() => {
+    const currentPath = router.pathname;
+    if (pathWithoutProtected.includes(currentPath)) return;
+
     const loginTimeStamp = mutateStorage.loginTimeStamp;
 
-    if (
-      !mutateStorage.accessToken ||
-      Date.now() - loginTimeStamp > 1000 * 60 * 60
-    ) {
+    if (!mutateStorage.accessToken || Date.now() - loginTimeStamp > 1000 * 60 * 60) {
       //TODO call sign-out
       mutateStorage.purge(); //clear storage on browser
-      router.push('/sign-in');
+      router.push("/sign-in");
     }
   }, []);
